[S12P31B106-309] 음성 채팅 진행상황 수정

See merge request s12-final/S12P31B106!305

+++ backend/src/main/java/com/voiceprint/backend/api/auth/dto/ProfileResponse.java
@@ -1,6 +1,5 @@
 package com.voiceprint.backend.api.auth.dto;
 
-import com.voiceprint.backend.domain.diary.Diary;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
@@ -11,7 +10,7 @@ import java.util.List;
 @Slf4j
 @AllArgsConstructor
 public class ProfileResponse {
-    private Long userId;
+    private Integer userId;
     private String nickname;
     private String imageUrl;
     private List<DiaryResponse> diaries;
